% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_str.R
\name{check_input_str}
\alias{check_input_str}
\title{check_input_str}
\usage{
check_input_str(.str)
}
\arguments{
\item{.str}{string}
}
\value{


Prepared string (IMPROVE)
}
\description{
check_input_str
}
\examples{
check_input_str('as.formula("G ~ A / B")')
check_input_str('as.formula(G ~ A / B)')
check_input_str('list("G ~ A / B")')
check_input_str('list(G ~ A / B)')
check_input_str('G ~ A / B')
}
